Nullkiller: fix file headers, move fuzzy config to ai folder
