+++ AI/VCAI/Behaviors/BuyArmyBehavior.cpp
@@ -1,5 +1,5 @@
 /*
-* Goals.cpp, part of VCMI engine
+* BuyArmyBehavior.cpp, part of VCMI engine
 *
 * Authors: listed in file AUTHORS in main folder
 *
