+++ AI/VCAI/Behaviors/CaptureObjectsBehavior.cpp
@@ -1,12 +1,12 @@
 /*
- * Goals.cpp, part of VCMI engine
- *
- * Authors: listed in file AUTHORS in main folder
- *
- * License: GNU General Public License v2.0 or later
- * Full text of license available in license.txt file, in main folder
- *
- */
+* CaptureObjectsBehavior.cpp, part of VCMI engine
+*
+* Authors: listed in file AUTHORS in main folder
+*
+* License: GNU General Public License v2.0 or later
+* Full text of license available in license.txt file, in main folder
+*
+*/
 #include "StdInc.h"
 #include "../VCAI.h"
 #include "../Engine/Nullkiller.h"
