+++ AI/VCAI/Engine/DangerHitMapAnalyzer.cpp
@@ -1,3 +1,12 @@
+/*
+* DangerHitMapAnalyzer.cpp, part of VCMI engine
+*
+* Authors: listed in file AUTHORS in main folder
+*
+* License: GNU General Public License v2.0 or later
+* Full text of license available in license.txt file, in main folder
+*
+*/
 #include "StdInc.h"
 #include "DangerHitMapAnalyzer.h"
