+++ AI/VCAI/Engine/DangerHitMapAnalyzer.h
@@ -1,3 +1,12 @@
+/*
+* DangerHitMapAnalyzer.h, part of VCMI engine
+*
+* Authors: listed in file AUTHORS in main folder
+*
+* License: GNU General Public License v2.0 or later
+* Full text of license available in license.txt file, in main folder
+*
+*/
 #pragma once
 
 #include "../VCAI.h"
