+++ AI/VCAI/Engine/Nullkiller.cpp
@@ -1,3 +1,12 @@
+/*
+* Nullkiller.cpp, part of VCMI engine
+*
+* Authors: listed in file AUTHORS in main folder
+*
+* License: GNU General Public License v2.0 or later
+* Full text of license available in license.txt file, in main folder
+*
+*/
 #include "StdInc.h"
 #include "Nullkiller.h"
 #include "../VCAI.h"
@@ -5,6 +14,7 @@
 #include "../Behaviors/CaptureObjectsBehavior.h"
 #include "../Behaviors/RecruitHeroBehavior.h"
 #include "../Behaviors/BuyArmyBehavior.h"
+#include "../Behaviors/StartupBehavior.h"
 #include "../Goals/Invalid.h"
 
 extern boost::thread_specific_ptr<CCallback> cb;
