+++ AI/VCAI/Engine/PriorityEvaluator.h
@@ -1,11 +1,11 @@
 /*
- * Fuzzy.h, part of VCMI engine
- *
- * Authors: listed in file AUTHORS in main folder
- *
- * License: GNU General Public License v2.0 or later
- * Full text of license available in license.txt file, in main folder
- *
+* PriorityEvaluator.h, part of VCMI engine
+*
+* Authors: listed in file AUTHORS in main folder
+*
+* License: GNU General Public License v2.0 or later
+* Full text of license available in license.txt file, in main folder
+*
 */
 #pragma once
 #include "fl/Headers.h"
