+++ AI/VCAI/Goals/ExecuteHeroChain.cpp
@@ -118,60 +118,3 @@ std::string ExecuteHeroChain::completeMessage() const
 {
 	return "Hero chain completed";
 }
-
-ExchangeSwapTownHeroes::ExchangeSwapTownHeroes(const CGTownInstance * town, const CGHeroInstance * garrisonHero)
-	:CGoal(Goals::EXCHANGE_SWAP_TOWN_HEROES), town(town), garrisonHero(garrisonHero)
-{
-}
-
-std::string ExchangeSwapTownHeroes::name() const
-{
-	return "Exchange and swap heroes of " + town->name;
-}
-
-std::string ExchangeSwapTownHeroes::completeMessage() const
-{
-	return "Exchange and swap heroes of " + town->name + " compelete";
-}
-
-bool ExchangeSwapTownHeroes::operator==(const ExchangeSwapTownHeroes & other) const
-{
-	return town == other.town;
-}
-
-TSubgoal ExchangeSwapTownHeroes::whatToDoToAchieve()
-{
-	return iAmElementar();
-}
-
-void ExchangeSwapTownHeroes::accept(VCAI * ai)
-{
-	if(!garrisonHero)
-	{
-		if(!town->garrisonHero)
-			throw cannotFulfillGoalException("Invalid configuration. There is no hero in town garrison.");
-		
-		cb->swapGarrisonHero(town);
-		ai->buildArmyIn(town);
-		ai->nullkiller->unlockHero(town->visitingHero.get());
-		logAi->debug("Extracted hero %s from garrison of %s", town->visitingHero->name, town->name);
-
-		return;
-	}
-
-	if(town->visitingHero && town->visitingHero.get() != garrisonHero)
-		cb->swapGarrisonHero(town);
-
-	ai->moveHeroToTile(town->visitablePos(), garrisonHero);
-
-	cb->swapGarrisonHero(town); // selected hero left in garrison with strongest army
-	ai->nullkiller->lockHero(town->garrisonHero.get());
-
-	if(town->visitingHero)
-	{
-		ai->nullkiller->unlockHero(town->visitingHero.get());
-		makePossibleUpgrades(town->visitingHero);
-	}
-
-	logAi->debug("Put hero %s to garrison of %s", town->garrisonHero->name, town->name);
-}
