+++ config/ai/object-priorities.txt
